Split tests for convert and stackstring
