+++ test/Jamfile.v2
@@ -10,32 +10,29 @@ project
     ;
     
    test-suite "nowide"
-        :   [ run test_convert.cpp ]
-            [ run test_fstream.cpp : : : <library>/boost/system//boost_system : test_fstream ]
-            [ run test_stdio.cpp ]
+        :   
             [ run test_codecvt.cpp ]
-            [ run test_env.cpp      :       : 
-                :   <toolset>msvc:<linkflags>shell32.lib : test_env_proto ]
-            [ run test_env.cpp      :       : 
-                :   <toolset>msvc:<linkflags>shell32.lib <define>BOOST_NOWIDE_TEST_INCLUDE_WINDOWS=1 : test_env_with_native ]
-                
-            [ run test_system.cpp   : "-w"  :
-                :   <toolset>msvc:<linkflags>shell32.lib : test_system_w ]
-                
-            [ run test_system.cpp   : "-n"  :
-                :   <toolset>msvc:<linkflags>shell32.lib : test_system_n ]
-                
-            [ run test_fs.cpp : : 
-                :   <library>/boost/filesystem//boost_filesystem
-                : test_fs ]
-
+            [ run test_convert.cpp ]
+            [ run test_env.cpp ]
+            [ run test_fstream.cpp ]
             [ run test_iostream.cpp : : 
                 :   <library>/boost/nowide//boost_nowide
                     <link>static 
                 : test_iostream_static ]
-                
             [ run test_iostream.cpp : : 
                 :   <library>/boost/nowide//boost_nowide
                     <link>shared
                 : test_iostream_shared ]
+            [ run test_stackstring.cpp ]
+            [ run test_stdio.cpp ]
+            [ run test_env.cpp : : 
+                :   <define>BOOST_NOWIDE_TEST_INCLUDE_WINDOWS=1 : test_env_win ]
+            [ run test_system.cpp : :
+                :   <define>NOWIDE_TEST_USE_NARROW=1 : test_system_n ]
+            [ run test_system.cpp : :
+                :   <define>NOWIDE_TEST_USE_NARROW=0 : test_system_w ]
+            [ run test_fs.cpp : : 
+                :   <library>/boost/filesystem//boost_filesystem
+                : test_fs ]
+
          ;
